% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox.R
\name{rox}
\alias{rox}
\title{rox: estimates concordance based on (R)ank (O)rder with (X)missing values}
\usage{
rox(xh, y, switching = T, w = NULL, ...)
}
\arguments{
\item{xh}{a vector of measurements that includes missing values, such as a metabolite measurement where mising values are due to left-truncation e.g. LOD}

\item{y}{a variable of interest that has no missing value, such as age and gender}

\item{switching}{TRUE by default. It is to decide whether switching to complete case analysis in the case of no evidence found for LOD assumption as described in buyukozkan et al.}

\item{w}{(optional) weight defined for missing values. By default it is estimated as described in buyukozkan et al.: w = 1 - (missingness ratio).}
}
\value{
estimated concordance index, standard error of it and corresponding p-value. Note that p-value is an approximate and not exact.
}
\description{
rox: estimates concordance based on (R)ank (O)rder with (X)missing values
}
\examples{
# sample size
n = 300
set.seed(42)
y = rnorm(n)
x =  y + rnorm(n)

# induce missing values
xh = x
xh[xh<0] = NA

# rox with missing values
rox::rox(xh, y)
# no missing values
rox::rox(x, y)

}
