% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_fit.R
\name{rox_mv}
\alias{rox_mv}
\title{multivaariable rox modeling}
\usage{
rox_mv(xh, ys, switching = T, w = NULL, func = c("concreg", "coxph"), ...)
}
\arguments{
\item{xh}{a vector of measurements that includes missing values, such as a metabolite measurement}

\item{ys}{a data.frame that includes variables of interest, no missing value is allowed}

\item{switching}{TRUE by default. It is to decide whether switching to complete case analysis in the case of no evidence found for LOD assumption as described in buyukozkan et al.}

\item{w}{(optional) weight defined for missing values. By default it is estimated as described in buyukozkan et al.: w = 1 - (missingness ratio).}

\item{func}{how to model multivariable relation, 'concreg'(default) is to use exponential link function to model
concordance probabilities. Another opiton is to use Cox's partial likelihood 'coxph' which is an approximate solution to 'concreg' but faster for big datasets.}
}
\value{
results and details of the fit
}
\description{
multivaariable rox modeling
}
\examples{

 # sample size for simulation
 n = 500

 set.seed(42)
 y = scale(rnorm(n))
 z = scale(rnorm(n))
 x = y + 2*z + rnorm(n)

 # induce missing values
 xh = x
 xh[ xh < -0.75] = NA

 # with missing values
 rox::rox_mv(xh, data.frame(z,y), func = "concreg")
 # no missing values
 rox::rox_mv(x, data.frame(z,y), func = "concreg")

}
